backend/pyproject.toml
[tool.poetry]
name = "tv-controller"
version = "0.1.0"
description = ""
authors = ["DACHXY <Danny10132024@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
pydantic = "^2.6.1"
pycaw = "^20230407"
comtypes = "^1.3.0"
pyalsaaudio = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
